Изменил Тарифы
TODO: Сделать тесты на тарифы
